move components to dir/file
