+++ src/HighTable.tsx
@@ -1,14 +1,14 @@
 import React, { MouseEvent, useCallback, useEffect, useMemo, useRef, useState } from 'react'
-import Cell from './components/Cell.js'
-import Row from './components/Row.js'
-import RowHeader from './components/RowHeader.js'
+import Cell from './components/Cell/Cell.js'
+import Row from './components/Row/Row.js'
+import RowHeader from './components/RowHeader/RowHeader.js'
+import TableHeader, { cellStyle } from './components/TableHeader/TableHeader.js'
 import { DataFrame } from './helpers/dataframe.js'
 import { PartialRow } from './helpers/row.js'
 import { Selection, SortIndex, areAllSelected, computeNewSelection, isSelected, toggleAll } from './helpers/selection.js'
 import { OrderBy, areEqualOrderBy } from './helpers/sort.js'
 import { rowError } from './helpers/text.js'
 import { useInputState } from './hooks/useInputState.js'
-import TableHeader, { cellStyle } from './TableHeader.js'
 import { stringify as stringifyDefault } from './utils/stringify.js'
 import { throttle } from './utils/throttle.js'
 export { arrayDataFrame, sortableDataFrame } from './helpers/dataframe.js'
